Remove the old middleware processors, pocessor trait, chain and builder, none of it is needed in the new middleware-as-layered-services world.
