+++ examples/serve-zone.rs
@@ -21,9 +21,9 @@ use domain::base::{Dname, Message, Rtype, ToDname};
 use domain::net::server::buf::VecBufSource;
 use domain::net::server::dgram::DgramServer;
 use domain::net::server::message::Request;
-use domain::net::server::middleware::processors::cookies_svc::CookiesMiddlewareSvc;
-use domain::net::server::middleware::processors::edns_svc::EdnsMiddlewareSvc;
-use domain::net::server::middleware::processors::mandatory_svc::MandatoryMiddlewareSvc;
+use domain::net::server::middleware::cookies::CookiesMiddlewareSvc;
+use domain::net::server::middleware::edns::EdnsMiddlewareSvc;
+use domain::net::server::middleware::mandatory::MandatoryMiddlewareSvc;
 use domain::net::server::service::{CallResult, ServiceError};
 use domain::net::server::stream::StreamServer;
 use domain::net::server::util::{mk_builder_for_target, service_fn};
